Fix list tag and space around slashes
I propose to use a narrow non-breaking space around the slashes. This provides good readability while not creating as large a gap as normal spaces would and at the same time prevents line breaks right before slashes so they'll never appear at the beginning of a new line but a rather at the end of the preceding.